@@ -1,83 +1,68 @@
 {
     "name": "tockler",
-    "version": "3.18.4",
     "private": true,
-    "dependencies": {
-        "@analytics/google-analytics": "^0.5.2",
-        "@react-hook/window-size": "^1.0.12",
-        "@welldone-software/why-did-you-render": "^3.5.0",
-        "analytics": "^0.7.5",
-        "antd": "^4.9.4",
-        "antd-theme-generator": "^1.2.8",
-        "antd-theme-webpack-plugin": "^1.3.7",
-        "babel-plugin-import": "^1.13.3",
-        "cross-env": "6.0.3",
-        "debounce-promise": "^3.1.2",
-        "easy-peasy": "^4.0.1",
-        "eiphop": "^1.0.13",
-        "electron-store": "5.2.0",
-        "emailjs-com": "^2.6.4",
-        "fast-deep-equal": "^3.1.1",
-        "less": "^3.13.1",
-        "less-loader": "^6.2.0",
-        "lodash": "4.17.19",
-        "loglevel": "^1.6.6",
-        "moment": "^2.24.0",
-        "moment-duration-format": "2.3.2",
-        "randomcolor": "^0.6.2",
-        "react": "^16.13.1",
-        "react-color": "2.19.3",
-        "react-component-queries": "2.3.0",
-        "react-dev-utils": "^10.0.0",
-        "react-dom": "^16.13.1",
-        "react-moment": "0.9.7",
-        "react-portal": "^4.2.0",
-        "react-router": "^5.1.2",
-        "react-router-dom": "^5.1.2",
-        "react-scripts": "^4.0.1",
-        "react-sizeme": "2.6.12",
-        "react-timeago": "4.4.0",
-        "react-use-form-state": "^0.13.2",
-        "reactcss": "^1.2.3",
-        "reflexbox": "^4.0.6",
-        "styled-components": "5.2.1",
-        "typeface-berkshire-swash": "0.0.72",
-        "use-react-router": "^1.0.7",
-        "victory": "^35.4.3"
-    },
+    "version": "0.0.0",
+    "type": "module",
     "scripts": {
-        "start": "cross-env ELECTRON_ENABLE_SECURITY_WARNINGS=false BROWSER=none LOG_LEVEL=debug react-app-rewired start",
-        "build": "cross-env GENERATE_SOURCEMAP=false PUBLIC_URL=./ react-app-rewired build",
-        "test": "yarn jest",
-        "eject": "react-app-rewired eject",
-        "lint": "eslint 'src/**/*{.ts,.tsx}' --fix"
+        "dev": "vite",
+        "build": "tsc -b && vite build",
+        "lint": "eslint .",
+        "test": "vitest",
+        "coverage": "vitest run --coverage",
+        "preview": "vite preview"
+    },
+    "dependencies": {
+        "@chakra-ui/icons": "^2.2.4",
+        "@chakra-ui/react": "^2.3.7",
+        "@chakra-ui/theme-tools": "^2.2.6",
+        "@datepicker-react/hooks": "^2.8.4",
+        "@emotion/react": "^11.10.5",
+        "@emotion/styled": "^11.10.5",
+        "@fontsource/inter": "^4.5.14",
+        "@hookform/resolvers": "^2.9.10",
+        "@react-hook/window-size": "^3.1.1",
+        "@testing-library/jest-dom": "^6.6.3",
+        "@testing-library/react": "^16.2.0",
+        "@vitest/coverage-v8": "^3.0.4",
+        "date-fns": "^4.1.0",
+        "debounce-promise": "^3.1.2",
+        "easy-peasy": "^5.1.0",
+        "fast-deep-equal": "^3.1.3",
+        "framer-motion": "^7",
+        "humanize-duration": "^3.32.1",
+        "lodash": "4.17.21",
+        "match-sorter": "^8.0.0",
+        "moment": "^2.30.1",
+        "randomcolor": "^0.6.2",
+        "react": "^19.0.0",
+        "react-color-palette": "^6.2.0",
+        "react-cool-dimensions": "^2.0.7",
+        "react-dom": "^19.0.0",
+        "react-hook-form": "^7.54.2",
+        "react-icons": "^5.4.0",
+        "react-moment": "1.1.3",
+        "react-portal": "^4.3.0",
+        "react-router-dom": "6.4.3",
+        "react-scripts": "^5.0.1",
+        "react-table": "^7.8.0",
+        "use-debounce": "^10.0.4",
+        "victory": "^36.6.8",
+        "vitest": "^3.0.4",
+        "yup": "^0.32.11"
     },
     "devDependencies": {
-        "@testing-library/jest-dom": "^4.2.4",
-        "@testing-library/react": "^9.3.2",
-        "@testing-library/user-event": "^7.1.2",
-        "@types/jest": "^24.0.0",
-        "@types/lodash.throttle": "^4.1.6",
-        "@types/node": "^12.0.0",
-        "@types/react": "^16.9.0",
-        "@types/react-dom": "^16.9.0",
-        "@types/webpack-env": "1.15.0",
-        "customize-cra": "^1.0.0",
-        "prettier": "^1.19.1",
-        "react-app-rewired": "^2.1.8",
-        "react-hot-loader": "4.12.19",
-        "ts-jest": "^26.0.0",
-        "typescript": "~3.7.2"
-    },
-    "eslintConfig": {
-        "extends": "react-app"
-    },
-    "browserslist": {
-        "production": [
-            "last 1 chrome version"
-        ],
-        "development": [
-            "last 1 chrome version"
-        ]
+        "@eslint/js": "^9.17.0",
+        "@types/react": "^19.0.8",
+        "@types/react-dom": "^19.0.3",
+        "@vitejs/plugin-react": "^4.3.4",
+        "babel-plugin-react-compiler": "19.0.0-beta-27714ef-20250124",
+        "eslint": "^9.17.0",
+        "eslint-plugin-react-compiler": "19.0.0-beta-27714ef-20250124",
+        "eslint-plugin-react-hooks": "^5.0.0",
+        "eslint-plugin-react-refresh": "^0.4.16",
+        "globals": "^15.14.0",
+        "typescript": "~5.6.2",
+        "typescript-eslint": "^8.18.2",
+        "vite": "^6.0.5"
     }
 }
