@@ -20,64 +20,59 @@
         "build_mac": "electron-builder -c electron-builder.yml --mac",
         "build_linux": "electron-builder -c electron-builder.yml --linux",
         "build_windows": "electron-builder -c electron-builder.yml --win",
-        "build": "webpack --progress --env.production --mode=production",
-        "prepare_client": "cpx \"../client/build/**/*\" dist",
+        "build": "webpack --progress --mode=production",
+        "prepare_client": "cpx \"../client/dist/**/*\" dist",
         "release": "electron-builder -c electron-builder.yml"
     },
-    "dependencies": {
-        "@sentry/electron": "^2.0.4",
-        "active-win": "^6.2.0",
-        "csv-stringify": "^5.5.3",
-        "eiphop": "^1.0.13",
-        "electron-context-menu": "2.3.0",
-        "electron-is-dev": "^1.2.0",
-        "electron-log": "^4.3.5",
-        "electron-store": "6.0.1",
-        "electron-updater": "4.3.5",
-        "hazardous": "^0.3.0",
-        "knex": "^0.21.13",
-        "knex-webpack-migration-source": "^1.0.0",
-        "lodash": "^4.17.20",
-        "menubar": "9.0.1",
-        "moment": "2.29.1",
-        "node-notifier": "^9.0.0",
-        "objection": "^2.2.3",
-        "randomcolor": "0.6.2",
-        "ref-napi": "^3.0.1",
-        "ref-struct-napi": "^1.1.1",
-        "sqlite3": "5.0.0"
+    "resolutions": {
+        "cliui": "^8.0.1"
     },
-    "peerDependencies": {},
+    "dependencies": {
+        "@sentry/electron": "^2.5.4",
+        "active-win": "^8.2.1",
+        "csv-stringify": "^6.2.0",
+        "electron-context-menu": "3.6.0",
+        "electron-is-dev": "^2.0.0",
+        "electron-log": "^4.4.8",
+        "electron-store": "8.1.0",
+        "electron-traywindow-positioner": "^1.2.1",
+        "electron-updater": "5.3.0",
+        "knex": "2.3.0",
+        "lodash": "^4.17.21",
+        "menubar": "9.5.1",
+        "moment": "2.30.1",
+        "node-machine-id": "^1.1.12",
+        "objection": "3.0.1",
+        "randomcolor": "0.6.2",
+        "sqlite3": "5.1.2",
+        "url-parse": "^1.5.10"
+    },
     "devDependencies": {
-        "@babel/register": "^7.12.10",
-        "@types/jest": "^26.0.19",
-        "@types/lodash": "^4.14.165",
-        "@types/node": "^14.14.14",
+        "@electron/notarize": "^2.5.0",
+        "@types/jest": "^29.2.2",
+        "@types/lodash": "^4.14.188",
+        "@types/node": "22.9.0",
+        "@types/randomcolor": "^0.5.9",
         "@types/webdriverio": "^5.0.0",
-        "@types/webpack": "^4.41.25",
-        "awesome-typescript-loader": "5.2.1",
-        "copy-webpack-plugin": "6.0.3",
+        "@types/webpack": "^5.28.5",
+        "@types/webpack-env": "^1.18.5",
+        "copy-webpack-plugin": "12.0.2",
         "cpx": "^1.5.0",
-        "cross-env": "^7.0.3",
-        "cucumber": "^6.0.5",
-        "cucumber-pretty": "^6.0.0",
-        "dotenv-webpack": "^6.0.0",
-        "electron": "7.3.2",
-        "electron-builder": "22.9.1",
-        "electron-connect": "0.6.3",
-        "electron-devtools-installer": "^3.1.1",
-        "electron-reload": "1.5.0",
-        "friendly-errors-webpack-plugin": "^1.7.0",
-        "jest": "26.6.3",
-        "jest-cli": "26.6.3",
-        "reflect-metadata": "^0.1.13",
-        "spectron": "^13.0.0",
-        "ts-jest": "^26.4.4",
-        "tslint": "^6.1.2",
-        "typescript": "^3.9.6",
-        "webpack": "^4.43.0",
-        "webpack-cli": "^3.3.12",
-        "webpack-node-externals": "^2.5.2"
+        "cross-env": "7.0.3",
+        "dotenv-webpack": "^8.1.0",
+        "electron": "16.2.8",
+        "electron-builder": "25.1.8",
+        "electron-devtools-installer": "^3.2.0",
+        "jest": "29.2.2",
+        "jest-cli": "29.2.2",
+        "node-gyp": "^10.2.0",
+        "ts-jest": "^29.0.3",
+        "ts-loader": "^9.5.1",
+        "tslint": "^6.1.3",
+        "typescript": "^5.6.3",
+        "webpack": "^5.96.1",
+        "webpack-cli": "5.1.4",
+        "webpack-node-externals": "^3.0.0"
     },
     "jest": {
         "globals": {
