@@ -1,22 +1,24 @@
-import React from 'react';
-import { Route, Switch } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
 import { SettingsPage } from './SettingsPage';
 import { SummaryPage } from './SummaryPage';
 import { TimelinePage } from './TimelinePage';
 import { TrayAppPage } from './TrayAppPage';
 import { SearchPage } from './SearchPage';
 import { SupportPage } from './SupportPage';
+import { MainLayout } from '../components/MainLayout/MainLayout';
 
-export function MainAppPage({ history }: any) {
+export function MainAppPage() {
     return (
-        <Switch>
-            <Route exact path="/" component={TimelinePage} />
-            <Route path="/app/timeline" component={TimelinePage} />
-            <Route path="/app/settings" component={SettingsPage} />
-            <Route path="/app/summary" component={SummaryPage} />
-            <Route path="/app/support" component={SupportPage} />
-            <Route path="/app/search" component={SearchPage} />
-            <Route path="/app/trayApp" component={TrayAppPage} />
-        </Switch>
+        <Routes>
+            <Route path="/" element={<MainLayout />}>
+                <Route index element={<TimelinePage />} />
+                <Route path="timeline" element={<TimelinePage />} />
+                <Route path="settings" element={<SettingsPage />} />
+                <Route path="summary" element={<SummaryPage />} />
+                <Route path="support" element={<SupportPage />} />
+                <Route path="search" element={<SearchPage />} />
+                <Route path="trayApp" element={<TrayAppPage />} />
+            </Route>
+        </Routes>
     );
 }
