@@ -1,22 +1,26 @@
-const path = require('path');
-const os = require('os');
+import CopyPlugin from 'copy-webpack-plugin';
+import Dotenv from 'dotenv-webpack';
+import nodeExternals from 'webpack-node-externals';
+import { fileURLToPath } from 'node:url';
+import path from 'node:path';
 
-const { TsConfigPathsPlugin, CheckerPlugin } = require('awesome-typescript-loader');
-const FriendlyErrorsWebpackPlugin = require('friendly-errors-webpack-plugin');
-const CopyPlugin = require('copy-webpack-plugin');
-const Dotenv = require('dotenv-webpack');
-const nodeExternals = require('webpack-node-externals');
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 
-module.exports = {
+export default {
     target: 'electron-main',
 
-    externals: [nodeExternals()],
+    externals: [
+        nodeExternals({
+            allowlist: ['get-windows'],
+        }),
+    ],
 
     resolve: {
         extensions: ['.ts', '.js'],
     },
     entry: {
-        index: path.resolve(__dirname, 'app', 'index.ts'),
+        index: path.resolve(__dirname, 'src', 'index.ts'),
     },
     output: {
         path: path.resolve(__dirname, 'dist'),
@@ -31,15 +35,14 @@ module.exports = {
         rules: [
             {
                 test: /\.ts$/i,
-                loader: 'awesome-typescript-loader',
+                loader: 'ts-loader',
+                options: {
+                    compilerOptions: {
+                        module: 'commonjs',
+                    },
+                },
             },
         ],
     },
-    plugins: [
-        new Dotenv(),
-        new FriendlyErrorsWebpackPlugin(),
-        new TsConfigPathsPlugin(),
-        new CheckerPlugin(),
-        new CopyPlugin({ patterns: ['preloadStuff.js'] }),
-    ],
+    plugins: [new Dotenv(), new CopyPlugin({ patterns: ['preloadStuff.js'] })],
 };
