@@ -1,3 +1,7 @@
-export const MODE_MONTH = 'month';
+export enum CALENDAR_MODE {
+    MONTH = 'month',
+    YEAR = 'year',
+}
+
 export const DAY_MONTH_FORMAT = 'MM-DD';
 export const MONTH_FORMAT = 'MM';
