@@ -1,23 +1,14 @@
 import _ from 'lodash';
-import {
-    convertDate,
-    TIME_FORMAT,
-    BREAKPOINT_TIME,
-    DURATION_FORMAT,
-    DURATION_SETTINGS,
-} from '../../constants';
+import { convertDate, TIME_FORMAT, BREAKPOINT_TIME } from '../../constants';
 import moment from 'moment';
-import { DAY_MONTH_FORMAT, MODE_MONTH } from '../../SummaryContext.util';
+import { DAY_MONTH_FORMAT, CALENDAR_MODE } from '../../SummaryContext.util';
 
-export const formatDuration = dur =>
-    moment.duration(dur).format(DURATION_FORMAT, DURATION_SETTINGS);
-
-export const groupByField = mode => item =>
-    mode === MODE_MONTH
+export const groupByField = (mode) => (item) =>
+    mode === CALENDAR_MODE.MONTH
         ? convertDate(item.beginDate).format(DAY_MONTH_FORMAT)
         : convertDate(item.beginDate).month();
 
-export const groupByActualDay = item => {
+export const groupByActualDay = (item) => {
     const date = convertDate(item.beginDate);
 
     if (date.format(TIME_FORMAT) < BREAKPOINT_TIME) {
@@ -32,7 +23,7 @@ export const summariseLog = (items, mode) => {
     _(items)
         .groupBy(groupByField(mode))
         .forEach((value, key) => {
-            data[key] = _.sumBy(value, c => convertDate(c.endDate).diff(convertDate(c.beginDate)));
+            data[key] = _.sumBy(value, (c) => convertDate(c.endDate).diff(convertDate(c.beginDate)));
         });
 
     return data;
@@ -42,10 +33,10 @@ export const summariseOnline = (items, mode) => {
     const data = {};
 
     _(items)
-        .filter(item => item.app === 'ONLINE')
+        .filter((item) => item.app === 'ONLINE')
         .groupBy(groupByField(mode))
         .forEach((value, key) => {
-            data[key] = _.sumBy(value, c => convertDate(c.endDate).diff(convertDate(c.beginDate)));
+            data[key] = _.sumBy(value, (c) => convertDate(c.endDate).diff(convertDate(c.beginDate)));
         });
     return data;
 };
@@ -58,13 +49,13 @@ export const summariseTimeOnline = (items, mode, beginDate) => {
     const currentMonth = beginDate.month();
 
     const data = _(items)
-        .filter(item => item.app === 'ONLINE')
+        .filter((item) => item.app === 'ONLINE')
         .groupBy(groupByActualDay)
-        .map((value, key) => {
+        .map((value) => {
             return {
-                beginDate: _.minBy(value, c => convertDate(c.beginDate)).beginDate,
-                endDate: _.maxBy(value, c => convertDate(c.endDate)).endDate,
-                online: _.sumBy(value, c => convertDate(c.endDate).diff(convertDate(c.beginDate))),
+                beginDate: _.minBy(value, (c) => convertDate(c.beginDate)).beginDate,
+                endDate: _.maxBy(value, (c) => convertDate(c.endDate)).endDate,
+                online: _.sumBy(value, (c) => convertDate(c.endDate).diff(convertDate(c.beginDate))),
             };
         })
         .reduce((result, currentValue) => {
