@@ -1,16 +1,34 @@
-import React, { useState } from 'react';
+import { useRef, useState } from 'react';
 import { Portal } from 'react-portal';
 import { Bar, VictoryTooltip } from 'victory';
+
+interface PropsI {
+    datum?: {};
+    onClickBarItem?: any;
+    getTooltipLabel: any;
+    popoverTriggerRef?: any;
+    x?: number;
+    y?: number;
+    theme: any;
+    centerTime?: boolean;
+}
+
+const calculateMiddle = (datum) => {
+    return datum.beginDate + (datum.endDate - datum.beginDate) / 2;
+};
 
 export const BarWithTooltip = ({
     datum = {},
     onClickBarItem,
     getTooltipLabel,
+    popoverTriggerRef = null,
     x = 0,
     y = 0,
     theme,
+    centerTime = false,
     ...rest
-}) => {
+}: PropsI) => {
+    const barRef = useRef(null);
     const [hover, setHover] = useState(false);
 
     const onMouseEnter = () => {
@@ -24,17 +42,31 @@ export const BarWithTooltip = ({
     const events = {
         onMouseEnter,
         onMouseLeave,
-        onClick: () => onClickBarItem(datum),
+        onClick: () => {
+            // Todo: Find a way to get barRef from Bar
+
+            if (onClickBarItem && popoverTriggerRef) {
+                console.info('barRef.current', barRef.current);
+                // eslint-disable-next-line react-compiler/react-compiler
+                popoverTriggerRef.current = barRef.current;
+                onClickBarItem(datum);
+            }
+        },
     };
+
+    const midpoint = calculateMiddle(datum);
+    // @ts-ignore
+    const newX = rest.scale.y(midpoint);
 
     return (
         <>
             {<Bar datum={datum} x={x} y={y} {...rest} events={events} />}
+
             {hover && (
                 <Portal closeOnEsc closeOnOutsideClick>
                     <VictoryTooltip
                         horizontal={false}
-                        x={x}
+                        x={centerTime ? newX : x}
                         y={y}
                         style={theme.tooltip.style}
                         cornerRadius={theme.tooltip.cornerRadius}
@@ -43,6 +75,7 @@ export const BarWithTooltip = ({
                         active
                         events={{}}
                         text={getTooltipLabel(datum)}
+                        datum={datum}
                     />
                 </Portal>
             )}
