@@ -1,97 +1,48 @@
-import { Menu, Tooltip } from 'antd';
-import { ClockCircleOutlined, PoweroffOutlined, ArrowsAltOutlined } from '@ant-design/icons';
-import React, { useState, useEffect, memo } from 'react';
-import tocklerIcon from '../../assets/icons/tockler_icon.png';
+import { memo } from 'react';
 import { EventEmitter } from '../../services/EventEmitter';
-import { getOnlineStartTime } from '../../services/trackItem.api';
-import { Brand, Img, MenuItem, RightMenuItem } from './TrayMenu.styles';
-import moment from 'moment';
-import Moment from 'react-moment';
-import { Logger } from '../../logger';
-import { useWindowFocused } from '../../hooks/windowFocusedHook';
-
-const getNow = () => moment().subtract(1, 'seconds');
+import { Box } from '@chakra-ui/react';
+import { AiOutlineArrowsAlt, AiOutlinePoweroff } from 'react-icons/ai';
+import { Tooltip } from '@chakra-ui/react';
+import { IconButton } from '@chakra-ui/react';
+import { HStack } from '@chakra-ui/react';
+import { Header } from '../Header/Header';
 
 export const TrayMenuPlain = () => {
-    const { windowIsActive } = useWindowFocused();
-    const [onlineSince, setOnlineSince] = useState();
     const exitApp = () => {
         EventEmitter.send('close-app');
     };
+
     const toggleMainWindow = () => {
         EventEmitter.send('toggle-main-window');
     };
 
-    useEffect(() => {
-        const systemIsOnline = () => {
-            Logger.debug('system-is-online');
-            setOnlineSince(getNow());
-        };
-        const systemIsNotOnline = () => {
-            Logger.debug('system-is-not-online');
-            setOnlineSince(null);
-        };
-
-        EventEmitter.on('system-is-online', systemIsOnline);
-        EventEmitter.on('system-is-not-online', systemIsNotOnline);
-
-        return () => {
-            EventEmitter.off('system-is-online', systemIsOnline);
-            EventEmitter.off('system-is-not-online', systemIsNotOnline);
-        };
-    }, []);
-
-    useEffect(() => {
-        if (windowIsActive) {
-            const loadOnlineStartTime = async () => {
-                const onlineStartTime = await getOnlineStartTime();
-
-                setOnlineSince(onlineStartTime ? moment(onlineStartTime) : getNow());
-            };
-
-            loadOnlineStartTime();
-        }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [windowIsActive]);
-
     return (
-        <Menu
-            mode="horizontal"
-            style={{ position: 'fixed', width: '100%', zIndex: 9000 }}
-            selectable={false}
-        >
-            <MenuItem key="/timeline2" onClick={toggleMainWindow}>
-                <Brand>
-                    <Img src={tocklerIcon} width="28" height="28" />
-                    <span>Tockler</span>
-                </Brand>
-            </MenuItem>
-            <MenuItem>
-                {onlineSince && (
-                    <Tooltip placement="bottom" title="Time without a break">
-                        <>
-                            <ClockCircleOutlined />{' '}
-                            <b>
-                                <Moment date={onlineSince} durationFromNow interval={60} />
-                            </b>
-                        </>
-                    </Tooltip>
-                )}
-            </MenuItem>
-            <RightMenuItem key="/exitApp" onClick={exitApp}>
-                <Tooltip placement="bottom" title="Quit app">
-                    <PoweroffOutlined />
+        <Header brandLinkProps={{ onClick: toggleMainWindow }}>
+            <Box flex="1"></Box>
+
+            <HStack spacing={3} pr={1}>
+                <Tooltip placement="bottom" label="Open main window">
+                    <IconButton
+                        onClick={toggleMainWindow}
+                        variant="ghost"
+                        colorScheme="gray"
+                        aria-label="Open main window"
+                        icon={<AiOutlineArrowsAlt />}
+                    />
                 </Tooltip>
-            </RightMenuItem>
-            <RightMenuItem key="/toggleMainWindow" onClick={toggleMainWindow}>
-                <Tooltip placement="bottom" title="Open main window">
-                    <ArrowsAltOutlined />
+
+                <Tooltip placement="bottom" label="Quit app">
+                    <IconButton
+                        onClick={exitApp}
+                        variant="ghost"
+                        colorScheme="gray"
+                        aria-label="Quit app"
+                        icon={<AiOutlinePoweroff />}
+                    />
                 </Tooltip>
-            </RightMenuItem>
-        </Menu>
+            </HStack>
+        </Header>
     );
 };
-
-TrayMenuPlain.whyDidYouRender = true;
 
 export const TrayMenu = memo(TrayMenuPlain);
