@@ -9,6 +9,8 @@ export const DURATION_FORMAT = 'w[w] d[d] h[h] m[m] s[s]';
 export const DURATION_SETTINGS = { largest: 2 };
 export const BREAKPOINT_TIME = '04:00';
 
-export const convertDate = (d: Date) => moment(d);
+export const convertDate = (d: Date) => {
+    return moment(d);
+};
 
 export const COLORS = { green: '#8BC34A' };
