@@ -1,7 +1,7 @@
 import { useEffect, useRef } from 'react';
 
-export function useInterval(callback, delay) {
-    const savedCallback = useRef<any>();
+export function useInterval(callback: () => void, delay: number | null): void {
+    const savedCallback = useRef<() => void>(() => {});
 
     // Remember the latest callback.
     useEffect(() => {
@@ -10,10 +10,10 @@ export function useInterval(callback, delay) {
 
     // Set up the interval.
     useEffect(() => {
-        function tick() {
+        function tick(): void {
             savedCallback.current();
         }
-        if (delay !== null) {
+        if (delay !== null && delay > 0) {
             const id = setInterval(tick, delay);
             return () => clearInterval(id);
         }
