@@ -1,11 +1,13 @@
-import React, { useState, createContext, useContext, useEffect } from 'react';
+import { useState, createContext, useContext, useEffect } from 'react';
 import { Logger } from '../logger';
 import { getChartTheme } from '../components/Timeline/ChartTheme';
 import { THEMES } from '../store/theme.util';
+import { useColorMode } from '@chakra-ui/react';
 
 export const ChartThemeContext = createContext({ chartTheme: getChartTheme(false) });
 
-export const ChartThemeProvider = ({ theme, children }) => {
+export const ChartThemeProvider = ({ children }) => {
+    const { colorMode } = useColorMode();
     const [chartTheme, setChartTheme] = useState(getChartTheme(false));
 
     const defaultContext = {
@@ -14,13 +16,11 @@ export const ChartThemeProvider = ({ theme, children }) => {
     };
 
     useEffect(() => {
-        Logger.info('Changing chartTheme');
+        Logger.info('Changing chartTheme', colorMode);
 
-        setChartTheme(theme.name === THEMES.DARK ? getChartTheme(true) : getChartTheme(false));
-    }, [theme]);
-    return (
-        <ChartThemeContext.Provider value={defaultContext}>{children}</ChartThemeContext.Provider>
-    );
+        setChartTheme(colorMode === THEMES.DARK ? getChartTheme(true) : getChartTheme(false));
+    }, [colorMode]);
+    return <ChartThemeContext.Provider value={defaultContext}>{children}</ChartThemeContext.Provider>;
 };
 
 export const useChartThemeState = () => {
