@@ -1,27 +1,54 @@
-import React, { useState } from 'react';
-import { MainLayout } from '../components/MainLayout/MainLayout';
-import { Box, Flex } from 'reflexbox';
-import emailjs from 'emailjs-com';
-import { Button, Input, Typography } from 'antd';
-const { Text, Title, Paragraph } = Typography;
+import { useState } from 'react';
+import { Button } from '@chakra-ui/react';
+import { Box, Flex, Text } from '@chakra-ui/react';
+import { Textarea } from '@chakra-ui/react';
+import { Input } from '@chakra-ui/react';
+import { CardBox } from '../components/CardBox';
+import { VStack } from '@chakra-ui/react';
 
-const TEMPLATE_ID = process.env.REACT_APP_TEMPLATE_ID || '';
-const SERVICE_ID = process.env.REACT_APP_SERVICE_ID || '';
-const USER_ID = process.env.REACT_APP_USER_ID || '';
+const TEMPLATE_ID = import.meta.env.VITE_TEMPLATE_ID || '';
+const SERVICE_ID = import.meta.env.VITE_SERVICE_ID || '';
+const USER_ID = import.meta.env.VITE_USER_ID || '';
 
-export function SupportPage({ location }: any) {
+const EMAILJS_API = 'https://api.emailjs.com/api/v1.0/email/send';
+
+const Paragraph = (props) => (
+    <Box py={2}>
+        <Text fontSize="lg" {...props} />
+    </Box>
+);
+
+const sendEmail = (templateParams) => {
+    var data = {
+        service_id: SERVICE_ID,
+        template_id: TEMPLATE_ID,
+        user_id: USER_ID,
+        template_params: { ...templateParams },
+    };
+
+    const requestOptions = {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify(data),
+    };
+    return fetch(EMAILJS_API, requestOptions);
+};
+
+const errToString = (err) => (err.text ? err.text : err.toString());
+
+export function SupportPage() {
     const [content, setContent] = useState('');
     const [contentError, setContentError] = useState(false);
     const [emailSent, setEmailSent] = useState(false);
     const [isSending, setIsSending] = useState(false);
     const [email, setEmail] = useState('');
 
-    const changeContent = e => {
+    const changeContent = (e) => {
         const { value } = e.target;
         setContent(value);
     };
 
-    const changeEmail = e => {
+    const changeEmail = (e) => {
         const { value } = e.target;
         setEmail(value);
     };
@@ -31,112 +58,50 @@ export function SupportPage({ location }: any) {
             setContentError(true);
             return;
         }
-        const templateParams = { reply_to: email, message: content };
+
         try {
             setIsSending(true);
-            await emailjs.send(SERVICE_ID, TEMPLATE_ID, templateParams, USER_ID);
+            await sendEmail({ reply_to: email, message: content });
             setEmailSent(true);
             setContent('');
             setContentError(false);
         } catch (err) {
-            alert('Email send error! ' + (err.text ? err.text : err.toString()));
+            alert('Email send error! ' + errToString(err));
         }
         setIsSending(false);
     };
 
     return (
-        <MainLayout location={location}>
-            <Box p={4} width={0.5}>
-                <Title level={3}>Contact Me</Title>
-                <Paragraph>
-                    Feel free to contact if you have any problems or feature requests. Or if you
-                    have any feedback to give - good or bad.
-                </Paragraph>
-                <Flex p={1}>
-                    <Input.TextArea
+        <VStack spacing={3} p={4} alignItems="flex-start">
+            <CardBox title="Contact Me" width={['100%', '100%', '100%', '100%', '50%']} divider>
+                <VStack spacing={3} alignItems="flex-start">
+                    <Paragraph>
+                        Feel free to contact if you have any problems or feature requests. Or if you have any feedback
+                        to give - good or bad.
+                    </Paragraph>
+
+                    <Textarea
+                        variant="outline"
                         value={content}
                         placeholder="Content"
                         onChange={changeContent}
                         rows={4}
                     />
-                </Flex>
-                {contentError && (
-                    <Flex p={1}>
-                        <Text>Content is empty!</Text>
-                    </Flex>
-                )}
-                <Flex p={1}>
-                    <Input
-                        value={email}
-                        placeholder="E-mail (If you need feedback)"
-                        onChange={changeEmail}
-                    />
-                </Flex>
-                <Flex p={1}>
-                    <Button type="primary" onClick={sendForm} disabled={isSending}>
+
+                    {contentError && <Text color="red">Content is empty!</Text>}
+
+                    <Input value={email} placeholder="E-mail (If you need feedback)" onChange={changeEmail} />
+
+                    <Button onClick={sendForm} disabled={isSending}>
                         Send
                     </Button>
-                </Flex>
+                </VStack>
                 {emailSent && (
-                    <Flex p={1}>
-                        <Text>Email is sent!</Text>
+                    <Flex pt={5}>
+                        <Text color="green">Email is sent!</Text>
                     </Flex>
                 )}
-            </Box>
-            <Box p={4} width={0.5}>
-                <Title level={3}>Support Tockler</Title>
-                <Paragraph>
-                    This app is made in my own free time and often at expense of family, friends,
-                    and sleep. I would like to keep this app free, open-source, and improving over
-                    time. But for that, your support is needed.
-                </Paragraph>
-                <Paragraph>
-                    It is understandable if you can't give anything. You can always give some
-                    constructive feedback.
-                </Paragraph>
-                <Paragraph>
-                    So if you find this app useful then feel free to donate. Anything helps to keep
-                    this app up to date and always improving. I don't plan to get rich with that,
-                    need to justify working on this.
-                </Paragraph>
-                <Flex p={1}>
-                    <Box p={2}>
-                        <a
-                            href="https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&amp;hosted_button_id=JAHHBZZCZVDMA"
-                            rel="noreferrer"
-                            target="_blank"
-                        >
-                            <img
-                                src="https://github.com/MayGo/tockler/raw/master/badges/Donate-PayPal-green.svg"
-                                alt="Donate using PayPal"
-                                style={{ maxWidth: '100%' }}
-                            />
-                        </a>
-                    </Box>
-                    <Box p={2}>
-                        <a
-                            href="https://github.com/sponsors/maygo/"
-                            rel="noreferrer"
-                            target="_blank"
-                        >
-                            <img
-                                src="https://github.com/MayGo/tockler/raw/master/badges/GitHub-Badge.svg"
-                                alt="Sponsor on GitHub"
-                                style={{ maxWidth: '100%' }}
-                            />
-                        </a>
-                    </Box>
-                    <Box p={2}>
-                        <a href="https://www.patreon.com/Tockler" rel="noreferrer" target="_blank">
-                            <img
-                                src="https://github.com/MayGo/tockler/raw/master/badges/Patreon-Badge.svg"
-                                alt="Become a patron"
-                                style={{ maxWidth: '100%' }}
-                            />
-                        </a>
-                    </Box>
-                </Flex>
-            </Box>
-        </MainLayout>
+            </CardBox>
+        </VStack>
     );
 }
