@@ -1,65 +1,133 @@
 import { EventEmitter } from './EventEmitter';
 import { Logger } from '../logger';
-import { emit } from 'eiphop';
 
-const { Store } = window as any;
+const { electronBridge } = window as any;
+const { configGet, configSet } = electronBridge;
 
-const config = new Store();
+const THEME_CONFIG_KEY = 'selectedTheme';
+const IS_NATIVE_THEME_ENABLED = 'isNativeThemeEnabled';
+const IS_LOGGING_ENABLED = 'isLoggingEnabled';
+const OPEN_AT_LOGIN = 'openAtLogin';
+const IS_AUTO_UPDATE_ENABLED = 'isAutoUpdateEnabled';
+const NATIVE_THEME_CONFIG_CHANGED = 'nativeThemeChanged';
+const USE_PURPLE_TRAY_ICON = 'usePurpleTrayIcon';
+const USE_PURPLE_TRAY_ICON_CHANGED = 'usePurpleTrayIconChanged';
+
+export function getNativeThemeChange() {
+    return configGet(IS_NATIVE_THEME_ENABLED) as boolean;
+}
+
+export function saveNativeThemeChange(enabled) {
+    configSet(IS_NATIVE_THEME_ENABLED, enabled);
+    EventEmitter.send(NATIVE_THEME_CONFIG_CHANGED);
+}
+
+// -----
 
 export function getOpenAtLogin() {
-    return config.get('openAtLogin') as boolean;
-}
-
-export function getIsAutoUpdateEnabled() {
-    return config.get('isAutoUpdateEnabled') as boolean;
-}
-export function getIsLoggingEnabled() {
-    return config.get('isLoggingEnabled') as boolean;
+    return configGet(OPEN_AT_LOGIN) as boolean;
 }
 
 export function saveOpenAtLogin(openAtLogin) {
     if (openAtLogin !== getOpenAtLogin()) {
+        configSet(OPEN_AT_LOGIN, openAtLogin);
         EventEmitter.send('openAtLoginChanged');
-        config.set('openAtLogin', openAtLogin);
     }
 }
 
-export function getThemeFromStorage() {
-    const activeTheme = config.get('activeTheme');
-    Logger.info('Got theme from storage:', activeTheme);
-    return activeTheme;
-}
+// -----
 
-export function saveThemeToStorage(theme) {
-    Logger.info('Save theme to storage:', theme);
-    config.set('activeTheme', theme);
+export function getIsAutoUpdateEnabled() {
+    return configGet(IS_AUTO_UPDATE_ENABLED) as boolean;
 }
 
 export function saveIsAutoUpdateEnabled(isAutoUpdateEnabled) {
     if (isAutoUpdateEnabled !== getIsAutoUpdateEnabled()) {
         Logger.debug('Setting isAutoUpdateEnabled', isAutoUpdateEnabled);
-        config.set('isAutoUpdateEnabled', isAutoUpdateEnabled);
+        configSet(IS_AUTO_UPDATE_ENABLED, isAutoUpdateEnabled);
     }
 }
+
+// -----
+
+export function getIsLoggingEnabled() {
+    return configGet(IS_LOGGING_ENABLED) as boolean;
+}
+
 export function saveIsLoggingEnabled(isLoggingEnabled) {
     if (isLoggingEnabled !== getIsLoggingEnabled()) {
         Logger.debug('Setting isLoggingEnabled', isLoggingEnabled);
-        config.set('isLoggingEnabled', isLoggingEnabled);
+        configSet(IS_LOGGING_ENABLED, isLoggingEnabled);
     }
+}
+
+// -----
+
+export function getUsePurpleTrayIcon() {
+    return configGet(USE_PURPLE_TRAY_ICON) as boolean;
+}
+
+export function saveUsePurpleTrayIcon(enabled) {
+    configSet(USE_PURPLE_TRAY_ICON, enabled);
+    EventEmitter.send(USE_PURPLE_TRAY_ICON_CHANGED);
+}
+
+export function getThemeFromStorage() {
+    const colorMode = configGet(THEME_CONFIG_KEY);
+    Logger.info('Got theme from storage:', colorMode);
+    return colorMode;
+}
+
+export function saveThemeToStorage(colorMode) {
+    Logger.info('Save theme to config:', colorMode);
+    return EventEmitter.emit('saveThemeAndNotify', colorMode);
 }
 
 export async function updateByName(name, jsonData) {
     Logger.debug('updateByName', JSON.stringify(jsonData));
+    return EventEmitter.emit('updateByName', { name, jsonData: JSON.stringify(jsonData) });
+}
 
-    return emit('updateByName', { name, jsonData: JSON.stringify(jsonData) });
+export async function notifyUser(message) {
+    Logger.debug('notifyUser', message);
+    return EventEmitter.emit('notifyUser', { message });
 }
 
 export function getRunningLogItem() {
-    return emit('getRunningLogItemAsJson');
+    return EventEmitter.emit('getRunningLogItemAsJson');
 }
 
-export function fetchWorkSettings() {
-    return emit('fetchWorkSettings');
+export function getMachineId() {
+    return EventEmitter.emit('getMachineId');
+}
+
+export async function fetchWorkSettings() {
+    const jsonStr = await EventEmitter.emit('fetchWorkSettingsJsonString');
+    try {
+        return JSON.parse(jsonStr);
+    } catch (e) {
+        Logger.error('Error in fetchWorkSettings', jsonStr, e);
+        return null;
+    }
+}
+
+export async function fetchDataSettings() {
+    const jsonStr = await EventEmitter.emit('fetchDataSettingsJsonString');
+    try {
+        return JSON.parse(jsonStr);
+    } catch (e) {
+        Logger.error('Error in fetchDataSettings', jsonStr, e);
+        return null;
+    }
+}
+
+export function saveWorkSettings(data) {
+    updateByName('WORK_SETTINGS', data);
+}
+
+export function saveDataSettings(data) {
+    updateByName('DATA_SETTINGS', data);
+    return EventEmitter.emit('updateByNameDataSettings', { name: 'DATA_SETTINGS', jsonData: JSON.stringify(data) });
 }
 
 export function saveAnalyserSettings(data) {
@@ -67,7 +135,7 @@ export function saveAnalyserSettings(data) {
 }
 
 export async function fetchAnalyserSettings() {
-    const jsonStr = await emit('fetchAnalyserSettingsJsonString');
+    const jsonStr = await EventEmitter.emit('fetchAnalyserSettingsJsonString');
     Logger.debug('fetchAnalyserSettings', jsonStr);
     try {
         return JSON.parse(jsonStr);
