@@ -3,15 +3,10 @@ import { ITrackItem } from '../@types/ITrackItem';
 import { TrackItemType } from '../enum/TrackItemType';
 import { Logger } from '../logger';
 import { EventEmitter } from './EventEmitter';
-import { emit } from 'eiphop';
 
-async function findAllDayItems(
-    from: moment.Moment,
-    to: moment.Moment,
-    taskName: string,
-): Promise<any> {
+async function findAllDayItems(from: moment.Moment, to: moment.Moment, taskName: string): Promise<any> {
     //Logger.debug('findAllDayItems', JSON.stringify({ from, to, taskName }));
-    const data = await emit('findAllDayItems', {
+    const data = await EventEmitter.emit('findAllDayItems', {
         from: from.valueOf(),
         to: to.valueOf(),
         taskName,
@@ -24,35 +19,37 @@ export async function findAllDayItemsForEveryTrack(from: moment.Moment, to: mome
     // TODO, query all at async
     const appItems: ITrackItem[] = await findAllDayItems(from, to, TrackItemType.AppTrackItem);
 
-    const statusItems: ITrackItem[] = await findAllDayItems(
-        from,
-        to,
-        TrackItemType.StatusTrackItem,
-    );
+    const statusItems: ITrackItem[] = await findAllDayItems(from, to, TrackItemType.StatusTrackItem);
     const logItems: ITrackItem[] = await findAllDayItems(from, to, TrackItemType.LogTrackItem);
 
     return { appItems, statusItems, logItems };
 }
 
 export function findFirstLogItems(): Promise<any> {
-    return emit('findFirstLogItems');
+    return EventEmitter.emit('findFirstLogItems');
+}
+
+export function findFirstTrackItem(): Promise<any> {
+    return EventEmitter.emit('findFirstTrackItem');
 }
 
 export function getOnlineStartTime(): Promise<any> {
-    return emit('getOnlineStartTime');
+    return EventEmitter.emit('getOnlineStartTime');
 }
 
-export function searchFromItems({ from, to, taskName, searchStr, paging }): Promise<any> {
-    return emit('searchFromItems', {
+export function searchFromItems({ from, to, taskName, searchStr, paging, sumTotal = false }): Promise<any> {
+    Logger.debug('Searching items:', { from, to, taskName, searchStr, paging });
+    return EventEmitter.emit('searchFromItems', {
         from: from.valueOf(),
         to: to.valueOf(),
         taskName,
         searchStr,
         paging,
+        sumTotal,
     });
 }
 export function exportFromItems({ from, to, taskName, searchStr }): Promise<any> {
-    return emit('exportFromItems', {
+    return EventEmitter.emit('exportFromItems', {
         from: from.valueOf(),
         to: to.valueOf(),
         taskName,
@@ -61,42 +58,57 @@ export function exportFromItems({ from, to, taskName, searchStr }): Promise<any>
 }
 
 function createTrackItem(trackItem: ITrackItem): Promise<any> {
-    return emit('createTrackItem', { trackItem: trackItem });
+    return EventEmitter.emit('createTrackItem', { trackItem: trackItem });
 }
 
-function updateTrackItem(trackItem: ITrackItem): Promise<any> {
-    return emit('updateTrackItem', { trackItem });
+function updateTrackItem(trackItem: ITrackItem, trackItemId: number): Promise<any> {
+    return EventEmitter.emit('updateTrackItem', { trackItem, trackItemId });
 }
 
-export async function saveTrackItem(trackItem): Promise<any> {
+function getRawTrackItem(savedItem) {
+    let item = {
+        id: savedItem.id,
+        app: savedItem.app,
+        title: savedItem.title,
+        taskName: savedItem.taskName,
+        color: savedItem.color,
+        beginDate: savedItem.beginDate,
+        url: savedItem.url,
+        endDate: savedItem.endDate,
+    };
+
+    return item;
+}
+
+export async function saveTrackItem(inputItem): Promise<any> {
+    const trackItem = getRawTrackItem(inputItem);
     Logger.debug('Saving trackitem.', trackItem);
+
     if (!trackItem.taskName) {
         trackItem.taskName = 'LogTrackItem';
     }
+
     if (trackItem.id) {
-        if (trackItem.originalColor === trackItem.color) {
-            // this.updateTrackItem(trackItem);
-        } else {
-            // this.showChangeColorDialog();
-        }
-        const item = await updateTrackItem(trackItem);
+        const item = await updateTrackItem(trackItem, trackItem.id);
         Logger.debug('Updated trackitem to DB:', item);
         return item;
     }
+
     if (!trackItem.app) {
         trackItem.app = 'Default';
     }
+
     const item = createTrackItem(trackItem);
     // Logger.debug('Created trackitem to DB:', item);
     return item;
 }
 
 export function deleteByIds(trackItemIds: number[]) {
-    return emit('deleteByIds', { trackItemIds });
+    return EventEmitter.emit('deleteByIds', { trackItemIds });
 }
 
 export function startNewLogItem(oldItem: any) {
-    Logger.debug('startNewLogItem');
+    Logger.debug('startNewLogItem', oldItem);
 
     const newItem: any = {};
     newItem.app = oldItem.app || 'WORK';
@@ -104,9 +116,7 @@ export function startNewLogItem(oldItem: any) {
     newItem.color = oldItem.color;
     newItem.title = oldItem.title;
     newItem.beginDate = moment().valueOf();
-    newItem.endDate = moment()
-        .add(60, 'seconds')
-        .valueOf();
+    newItem.endDate = moment().add(60, 'seconds').valueOf();
 
     EventEmitter.send('start-new-log-item', newItem);
 }
@@ -117,5 +127,5 @@ export function stopRunningLogItem(runningLogItemId: number) {
 }
 
 export function updateTrackItemColor(appName: string, color: string) {
-    return emit('updateTrackItemColor', { appName, color });
+    return EventEmitter.emit('updateTrackItemColor', { appName, color });
 }
