@@ -1,22 +1,20 @@
-import { Application, SpectronClient } from "spectron";
-import { join } from "path";
+import { Application } from 'spectron';
 import * as utils from './utils';
 
 type TestContext = {
-    app: Application
+    app: Application;
 };
 function sleep(ms) {
-    return new Promise(resolve => setTimeout(resolve, ms));
+    return new Promise((resolve) => setTimeout(resolve, ms));
 }
 let app: Application;
-describe("application launch", function() {
 
+describe('application launch', function () {
     beforeEach(async () => {
         app = utils.createApplication();
         await utils.startApplication(app);
         app.client.windowByIndex(1);
         await app.client.waitUntilWindowLoaded();
-
     });
 
     afterEach(async () => {
@@ -27,5 +25,4 @@ describe("application launch", function() {
         const text = await app.client.getText('#appName');
         expect(text).toEqual('Tockler');
     });
-
 });
