@@ -1,5 +1,5 @@
 import { TrackItemType } from '../app/enums/track-item-type';
-import * as moment from 'moment';
+import moment from 'moment';
 
 export default class TrackItemTestData {
     static getAppTrackItem(data = {}, addHours = 0): any {
@@ -8,10 +8,7 @@ export default class TrackItemTestData {
             title: 'google.com',
             taskName: TrackItemType.AppTrackItem,
             color: '#123456',
-            beginDate: moment()
-                .startOf('day')
-                .add(addHours, 'hours')
-                .toDate(),
+            beginDate: moment().startOf('day').add(addHours, 'hours').toDate(),
             endDate: moment()
                 .startOf('day')
                 .add(addHours + 1, 'hours')
@@ -27,10 +24,7 @@ export default class TrackItemTestData {
             title: 'Some work.',
             taskName: TrackItemType.LogTrackItem,
             color: '#123456',
-            beginDate: moment()
-                .startOf('day')
-                .add(addDays, 'hours')
-                .toDate(),
+            beginDate: moment().startOf('day').add(addDays, 'hours').toDate(),
             endDate: moment()
                 .startOf('day')
                 .add(addDays + 1, 'hours')
@@ -45,18 +39,13 @@ export default class TrackItemTestData {
             app: 'IDLE',
             taskName: TrackItemType.StatusTrackItem,
             color: '#123456',
-            beginDate: moment()
-                .startOf('day')
-                .add(addDays, 'hours')
-                .toDate(),
+            beginDate: moment().startOf('day').add(addDays, 'hours').toDate(),
             endDate: moment()
                 .startOf('day')
                 .add(addDays + 1, 'hours')
                 .toDate(),
         };
-        rawItem.title = data.app
-            ? data.app.toString().toLowerCase()
-            : rawItem.app.toString().toLowerCase();
+        rawItem.title = data.app ? data.app.toString().toLowerCase() : rawItem.app.toString().toLowerCase();
         return Object.assign({}, rawItem, data);
     }
 
@@ -65,18 +54,13 @@ export default class TrackItemTestData {
             app: 'ONLINE',
             taskName: TrackItemType.StatusTrackItem,
             color: '#666777',
-            beginDate: moment()
-                .startOf('day')
-                .add(addDays, 'hours')
-                .toDate(),
+            beginDate: moment().startOf('day').add(addDays, 'hours').toDate(),
             endDate: moment()
                 .startOf('day')
                 .add(addDays + 1, 'hours')
                 .toDate(),
         };
-        rawItem.title = data.app
-            ? data.app.toString().toLowerCase()
-            : rawItem.app.toString().toLowerCase();
+        rawItem.title = data.app ? data.app.toString().toLowerCase() : rawItem.app.toString().toLowerCase();
         return Object.assign({}, rawItem, data);
     }
 }
