@@ -1,9 +1,9 @@
-exports.up = function(knex) {
+exports.up = function (knex) {
     // We renamed table TrackItem to TrackItems at some point with sequalize.
     // Try to recover from that
-    const trackItemsTable = knex.schema.hasTable('TrackItem').then(function(oldExists) {
+    const trackItemsTable = knex.schema.hasTable('TrackItem').then(function (oldExists) {
         if (oldExists) {
-            return knex.schema.hasTable('TrackItem').then(function(newExists) {
+            return knex.schema.hasTable('TrackItem').then(function (newExists) {
                 if (newExists) {
                     // If we have old table TrackItem and new TrackItems then delete old table
                     return knex.schema.dropTable('TrackItem');
@@ -13,9 +13,9 @@ exports.up = function(knex) {
                 }
             });
         } else {
-            return knex.schema.hasTable('TrackItems').then(function(newExists) {
+            return knex.schema.hasTable('TrackItems').then(function (newExists) {
                 if (!newExists) {
-                    return knex.schema.createTable('TrackItems', function(table) {
+                    return knex.schema.createTable('TrackItems', function (table) {
                         table.increments('id');
                         table.string('app', 255);
                         table.string('taskName', 255);
@@ -29,9 +29,9 @@ exports.up = function(knex) {
         }
     });
 
-    const appSettingTable = knex.schema.hasTable('AppSetting').then(function(oldExists) {
+    const appSettingTable = knex.schema.hasTable('AppSetting').then(function (oldExists) {
         if (oldExists) {
-            return knex.schema.hasTable('AppSettings').then(function(newExists) {
+            return knex.schema.hasTable('AppSettings').then(function (newExists) {
                 if (newExists) {
                     // If we have old table AppSetting and new AppSettings then delete old table
                     return knex.schema.dropTable('AppSetting');
@@ -41,23 +41,33 @@ exports.up = function(knex) {
                 }
             });
         } else {
-            return knex.schema.hasTable('AppSettings').then(function(newExists) {
+            return knex.schema.hasTable('AppSettings').then(function (newExists) {
                 if (!newExists) {
-                    return knex.schema.createTable('AppSettings', function(table) {
-                        table.increments('id');
-                        table.string('name', 255);
-                        table.string('color', 255);
-                    });
+                    return knex.schema
+                        .createTable('AppSettings', function (table) {
+                            table.increments('id');
+                            table.string('name', 255);
+                            table.string('color', 255);
+                        })
+                        .then(function () {
+                            return knex('AppSettings').insert([
+                                { name: 'ONLINE', color: '#7ed321' },
+                                { name: 'OFFLINE', color: '#f31b1b' },
+                                { name: 'IDLE', color: '#f5a623' },
+                            ]);
+                        });
                 }
             });
         }
     });
 
-    return Promise.all([trackItemsTable, appSettingTable].map(p => p.catch(e => console.error(e))))
-        .then(results =>
-            knex.schema.hasTable('Settings').then(function(newExists) {
+    return Promise.all(
+        [trackItemsTable, appSettingTable].map((p) => p.catch((e) => console.error(e))),
+    )
+        .then((results) =>
+            knex.schema.hasTable('Settings').then(function (newExists) {
                 if (!newExists) {
-                    return knex.schema.createTable('Settings', function(table) {
+                    return knex.schema.createTable('Settings', function (table) {
                         table.increments('id');
                         table.string('name', 255);
                         table.string('jsonData', 255);
@@ -65,12 +75,9 @@ exports.up = function(knex) {
                 }
             }),
         )
-        .catch(e => console.log(e));
+        .catch((e) => console.log(e));
 };
 
-exports.down = function(knex) {
-    return knex.schema
-        .dropTable('TrackItems')
-        .dropTable('AppSettings')
-        .dropTable('Settings');
+exports.down = function (knex) {
+    return knex.schema.dropTable('TrackItems').dropTable('AppSettings').dropTable('Settings');
 };
